Refactor Decompiler
Remove the strange behavior where we check the first byte of the
compiled program. I wasn't able to find a compiler that replicated the
behavior of writing the first byte as a '2' instead of '2a', and the two
major compilers that I assume people use (Tokens and SourceCoder),
including dallas, all write the first byte as '2a', and all successive
bytes in such a fashion.

Stop cleaning the program before we begin compilation. Binary files
don't contain newlines or spaces, so we don't have to sanitize it before
we can begin. Because we aren't cleaning anymore, this fixed the problem
where 2 had to be subtracted from the dataOffset and titleOffset.

Stop title decompilation if we find a 0x00 early.

Rename split() to combine2ByteElements().

backwardsLex() now returns an error along with the token slice.

Add 0x63 to the list of 2-byte delimeters.