Add framework for generating token map
Not sure if this will get used, but I'm trying to move away from the
dual-map scenario that currently exists (one for single byte tokens and
one for two byte tokens). In order to do this, I've been thinking about
using strings instead of integers to represent the tokens, which will
require a rework of the compiler in order to write the output file as
bytes instead of the strings that the map holds.